Move new ES client and readed impl to the main package
