+++ pkg/esclient/client_test.go
