+++ pkg/esclient/es6client_test.go
