+++ pkg/esclient/ping.go
