+++ pkg/esclient/ping_test.go
