+++ plugin/storage/esnew/dependencystore/dependency_store.go
@@ -23,8 +23,8 @@ import (
 
 	"go.uber.org/zap"
 
-	"github.com/jaegertracing/jaeger/cmd/opentelemetry/app/exporter/elasticsearchexporter/esclient"
 	"github.com/jaegertracing/jaeger/model"
+	"github.com/jaegertracing/jaeger/pkg/esclient"
 	"github.com/jaegertracing/jaeger/plugin/storage/es/dependencystore/dbmodel"
 	"github.com/jaegertracing/jaeger/storage/dependencystore"
 )
