+++ plugin/storage/esnew/spanreader/validate_query.go
